+++ packages/mask/src/extension/popups/pages/Personas/AccountDetail/index.tsx
@@ -8,6 +8,7 @@ import {
     SignType,
     SOCIAL_MEDIA_SUPPORTING_NEXT_DOT_ID,
 } from '@masknet/shared-base'
+import { PersonaContext } from '@masknet/shared'
 import { NextIDProof } from '@masknet/web3-providers'
 import { usePopupCustomSnackbar } from '@masknet/theme'
 import { useTitle } from '../../../hook/useTitle.js'
@@ -15,7 +16,6 @@ import { useI18N } from '../../../../../utils/index.js'
 import { AccountDetailUI } from './UI.js'
 import Service from '../../../../service.js'
 import { DisconnectDialog } from '../components/DisconnectDialog/index.js'
-import { PersonaContext } from '@masknet/shared'
 
 const AccountDetail = memo(() => {
     const { t } = useI18N()
