+++ packages/mask/src/extension/popups/pages/Personas/components/PersonaHeader/index.tsx
@@ -1,10 +1,10 @@
 import { memo } from 'react'
 import { useMatch, useNavigate } from 'react-router-dom'
+import { PersonaContext } from '@masknet/shared'
 import { PopupRoutes } from '@masknet/shared-base'
 import { PersonaHeaderUI } from './UI.js'
 import { NormalHeader } from '../../../../components/NormalHeader/index.js'
 import Services from '../../../../../service.js'
-import { PersonaContext } from '@masknet/shared'
 
 export const PersonaHeader = memo(() => {
     const navigate = useNavigate()
