+++ packages/plugins/FileService/src/SNSAdaptor/contexts/FileManagement/index.tsx
@@ -1,5 +1,3 @@
-import type { CompositionType } from '@masknet/plugin-infra/content-script'
-import { CrossIsolationMessages, EMPTY_LIST } from '@masknet/shared-base'
 import { noop, omit } from 'lodash-es'
 import {
     createContext,
@@ -12,13 +10,15 @@ import {
     useMemo,
     useState,
 } from 'react'
-import { useNavigate } from 'react-router-dom'
 import { useAsyncRetry } from 'react-use'
+import { useNavigate } from 'react-router-dom'
+import type { CompositionType } from '@masknet/plugin-infra/content-script'
+import { ApplicationBoardDialog } from '@masknet/shared'
+import { CrossIsolationMessages, EMPTY_LIST } from '@masknet/shared-base'
 import { META_KEY_3, RoutePaths } from '../../../constants.js'
 import { digest, makeFileKey } from '../../../helpers.js'
 import type { FileInfo, Provider } from '../../../types.js'
 import { PluginFileServiceRPC, PluginFileServiceRPCGenerator } from '../../../Worker/rpc.js'
-import { ApplicationBoardDialog } from '@masknet/shared'
 
 interface UploadState {
     progress: number
