+++ packages/shared/src/hooks/index.ts
@@ -6,6 +6,7 @@ export * from './useImageURL.js'
 export * from './useIsImageURL.js'
 export * from './useTokenSecurity.js'
 export * from './useSocialAccountsBySettings.js'
+export * from './usePersonaContext.js'
 export * from './usePersonasFromNextID.js'
 export * from './usePersonaConnectStatus.js'
 export * from './usePersonaProofs.js'
@@ -15,4 +16,3 @@ export * from './useTokenMenuCollectionList.js'
 export * from './useAvailableBalance.js'
 export * from './useSingletonModal.js'
 export * from './useMatchXS.js'
-export * from './usePersonaContext/index.js'
